+++ test/smokes/golint/expectations.rb
@@ -21,11 +21,12 @@ Smoke.add_test(
   },
   warnings: [
     {
-      message: <<~MSG.strip,
+      message: <<~MSG
         DEPRECATION WARNING!!!
         The support for golint is deprecated. Sider will drop these versions on March 31, 2020.
-        Please consider using an alternative tool GolangCi-Lint.
+        Please consider using an alternative tool GolangCi-Lint. See https://help.sider.review/tools/go/golint
       MSG
+        .strip,
       file: "sider.yml"
     }
   ]