+++ test/smokes/gometalinter/expectations.rb
@@ -7,11 +7,12 @@ Smoke.add_test(
   },
   warnings: [
     {
-      message: <<~MSG.strip,
-          DEPRECATION WARNING!!!
-          The support for gometalinter is deprecated. Sider will drop these versions on March 31, 2020.
-          Please consider using an alternative tool GolangCi-Lint.
+      message: <<~MSG
+        DEPRECATION WARNING!!!
+        The support for gometalinter is deprecated. Sider will drop these versions on March 31, 2020.
+        Please consider using an alternative tool GolangCi-Lint. See https://github.com/alecthomas/gometalinter/issues/590
       MSG
+        .strip,
       file: "sider.yml"
     }
   ]
@@ -36,16 +37,7 @@ Smoke.add_test(
     ],
     analyzer: { name: "gometalinter", version: "2.0.11" }
   },
-  warnings: [
-    {
-      message: <<~MSG.strip,
-          DEPRECATION WARNING!!!
-          The support for gometalinter is deprecated. Sider will drop these versions on March 31, 2020.
-          Please consider using an alternative tool GolangCi-Lint.
-      MSG
-      file: "sideci.yml"
-    }
-  ]
+  warnings: [{ message: /The support for gometalinter is deprecated/, file: "sideci.yml" }]
 )
 
 Smoke.add_test(
@@ -67,16 +59,7 @@ Smoke.add_test(
     ],
     analyzer: { name: "gometalinter", version: "2.0.11" }
   },
-  warnings: [
-    {
-      message: <<~MSG.strip,
-          DEPRECATION WARNING!!!
-          The support for gometalinter is deprecated. Sider will drop these versions on March 31, 2020.
-          Please consider using an alternative tool GolangCi-Lint.
-      MSG
-      file: "sideci.yml"
-    }
-  ]
+  warnings: [{ message: /The support for gometalinter is deprecated/, file: "sideci.yml" }]
 ) { |config| config.ssh_key = "ssh_key" }
 
 Smoke.add_test(
@@ -135,16 +118,7 @@ Smoke.add_test(
     ],
     analyzer: { name: "gometalinter", version: "2.0.11" }
   },
-  warnings: [
-    {
-      message: <<~MSG.strip,
-        DEPRECATION WARNING!!!
-        The support for gometalinter is deprecated. Sider will drop these versions on March 31, 2020.
-        Please consider using an alternative tool GolangCi-Lint.
-      MSG
-      file: "sideci.yml"
-    }
-  ]
+  warnings: [{ message: /The support for gometalinter is deprecated/, file: "sideci.yml" }]
 )
 
 Smoke.add_test(
@@ -153,15 +127,8 @@ Smoke.add_test(
     guid: "test-guid", timestamp: :_, type: "success", issues: [], analyzer: { name: "gometalinter", version: "2.0.11" }
   },
   warnings: [
-    { message: "`install_path` option is deprecated. Use `import_path` instead.", file: "sideci.yml" },
-    {
-      message: <<~MSG.strip,
-        DEPRECATION WARNING!!!
-        The support for gometalinter is deprecated. Sider will drop these versions on March 31, 2020.
-        Please consider using an alternative tool GolangCi-Lint.
-      MSG
-      file: "sideci.yml"
-    }
+    { message: /The support for gometalinter is deprecated/, file: "sideci.yml" },
+    { message: "`install_path` option is deprecated. Use `import_path` instead.", file: "sideci.yml" }
   ]
 )
 
@@ -170,16 +137,7 @@ Smoke.add_test(
   {
     guid: "test-guid", timestamp: :_, type: "success", issues: [], analyzer: { name: "gometalinter", version: "2.0.11" }
   },
-  warnings: [
-    {
-      message: <<~MSG.strip,
-          DEPRECATION WARNING!!!
-          The support for gometalinter is deprecated. Sider will drop these versions on March 31, 2020.
-          Please consider using an alternative tool GolangCi-Lint.
-      MSG
-      file: "sideci.yml"
-    }
-  ]
+  warnings: [{ message: /The support for gometalinter is deprecated/, file: "sideci.yml" }]
 )
 
 Smoke.add_test(
@@ -187,14 +145,5 @@ Smoke.add_test(
   {
     guid: "test-guid", timestamp: :_, type: "success", issues: [], analyzer: { name: "gometalinter", version: "2.0.11" }
   },
-  warnings: [
-    {
-      message: <<~MSG.strip,
-          DEPRECATION WARNING!!!
-          The support for gometalinter is deprecated. Sider will drop these versions on March 31, 2020.
-          Please consider using an alternative tool GolangCi-Lint.
-      MSG
-      file: "sideci.yml"
-    }
-  ]
+  warnings: [{ message: /The support for gometalinter is deprecated/, file: "sideci.yml" }]
 )